Refactor DBBSC ECMD modules
